Updated Berthing Class to instead be a generic Room Class that accepts a Description and a Map that holds the exits, adjusted testing file ,AB#202
